Step by step, refactoring....
