+++ lib/pandora/constants.rb
@@ -1,0 +1,39 @@
+module Pandora
+  module Constants
+    # Property indexes of field definition array
+    # RU: Индексы свойств в массиве описания полей
+    FI_Id      = 0
+    FI_Name    = 1
+    FI_Type    = 2
+    FI_Size    = 3
+    FI_Pos     = 4
+    FI_FSize   = 5
+    FI_Hash    = 6
+    FI_View    = 7
+    FI_LName   = 8
+    FI_VFName  = 9
+    FI_Index   = 10
+    FI_LabOr   = 11
+    FI_NewRow  = 12
+    FI_VFSize  = 13
+    FI_Value   = 14
+    FI_Widget  = 15
+    FI_Label   = 16
+    FI_LabW    = 17
+    FI_LabH    = 18
+    FI_WidW    = 19
+    FI_WidH    = 20
+    FI_Color   = 21
+    FI_Widget2 = 22
+
+    PF_Name    = 0
+    PF_Desc    = 1
+    PF_Type    = 2
+    PF_Section = 3
+    PF_Setting = 4
+
+    TI_Name  = 0
+    TI_Type  = 1
+    TI_Desc  = 2
+  end
+end
