+++ lib/pandora/model/model.rb
@@ -1,0 +1,12 @@
+module Pandora
+  module Model
+
+    # Pandora's object
+    # RU: Объект Пандоры
+    class Panobject < Pandora::Utils::BasePanobject
+      ider = 'Panobject'
+      name = "Объект Пандоры"
+    end
+
+  end
+end
